Got magic to generate

+++ crates/engine/build/build.rs
@@ -1,8 +1,8 @@
 #![feature(generic_const_exprs)]
 #![feature(associated_type_defaults)]
+#![feature(const_trait_impl)]
 
 use std::{fs::File, io::Write, path::Path};
-
 
 #[path = "../src/bitboard/mod.rs"]
 mod bitboard;
@@ -23,7 +23,7 @@ fn main() {
         File::create(
             Path::new("/home/zioth/projects/apps/chess/crates/engine/data/built_at")
                 .join(format!("{:?}.txt", time).replace("\"", ""))
-/*   */       )
+        )
         .unwrap(),
         "{:?}",
         now,
@@ -33,7 +33,9 @@ fn main() {
     // let out_dir = env::var("OUT_DIR").unwrap();
     let out_dir = "/home/zioth/projects/apps/chess/crates/engine/data";
     let magic_path = Path::new(&out_dir).join("magic_gen.rs");
+    if magic_path.metadata().is_ok_and(|f| f.is_file()) { return; }
     let mut file = File::create(&magic_path).unwrap();
+
 
     generators::LinesGenerator::write_generated_array(&mut file).unwrap();
     generators::BetweenGenerator::write_generated_array(&mut file).unwrap();
@@ -44,6 +46,7 @@ fn main() {
     generators::PawnAttacksGenerator::write_generated_array(&mut file).unwrap();
     generators::PawnDoubleMoveOriginsGenerator::write_generated_value(&mut file).unwrap();
     generators::PawnDoubleMoveDestinationsGenerator::write_generated_value(&mut file).unwrap();
+    generators::MagicGenerator::write_generated_array(&mut file).unwrap();
 
     // println!("BUILD SCRIPT RUNNING");
 }
